Support "newstyle" Youtube playlist IDs.
Many playlists reported by Youtube now have the form like in:

    http://www.youtube.com/playlist?list=PL520044A3524F5E5D

where `PL` is prefixed to what youtube_search used to use as playlist IDs.
So, we discard the `PL` with a regular expression and only get the `520044A3524F5E5D` in the case of the playlist of the
example.